INS-1946: removed unused functions, moved MergedListCopy from interfaces.go (and all functions that work with it) to network/nodenetwork/utils.go
